LP-74 Refactor packages
